% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipper.r
\name{clipper}
\alias{clipper}
\title{A clip function to create an outliner for national or state maps}
\usage{
clipper(geoobj, state, margin = 0.1)
}
\arguments{
\item{geoobj}{A geo-object,}

\item{state}{a state abbrebriation, e.g. "CA",}

\item{margin}{margin added to maps, %}

\item{yName}{yName for making equation labels}
}
\description{
This function performed basic clip function to geoobj and add a polygon outside line
}
\examples{
clip1 <- clipper(geoobj = us.states, state = unique(us.states$STATE_ABBR))  # USA map with  projection
plot(clip1,col = "red", border ="black", xpd =F)  
clip2 <- clipper(geoobj =  us.states, state = "CA")   # CA map with nhd2 projection
plot(clip2, col ="blue")
clip3 <- clipper(geoobj = us.counties, state = as.vector(unique(us.counties$STATE_ABBR)))  # USA map with  projection
plot(clip3,col = "red", border ="black", xpd =F) 
}
