% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colbar.r
\name{colbar}
\alias{colbar}
\title{A function to plot US map and add colors to each state based on values}
\usage{
colbar(geoobj = us.states, values = runif(nrow(us.states) + 1, 0, 100),
  type = "rgb", ltitle = "Density", unit = "/mile", main = "Population",
  scheme = c("purple4", "blue4", "blue1", "green3", "yellowgreen", "yellow2",
  "sandybrown", "red", "darkred"))
}
\arguments{
\item{geoobj}{A geo-object,}

\item{values}{assigned values for each state to create color gradient,}

\item{type}{seven types to pick from c("rgb","topo","heat", "rain","cm","ter"(rain), "gray")}

\item{ltitle}{legend title}

\item{unit}{unit for legend}

\item{scheme}{color scheme for rgb.palette, only used for rbg.palette}
}
\description{
This function used a stored US state and county GIS data to plot color maps to each state and county;
the US map is projected to Albers
}
\examples{
colbar(geoobj=us.states, values= runif(nrow(us.states)+1, 0,100), # 1:51, #
type = "rgb.col", ltitle ="Density",
unit ="/mile", main ="Population",
scheme = c("purple4", "blue4","blue1","green3", "yellowgreen","yellow2","sandybrown","red","darkred") )
}
