% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapscale.r
\name{mapscale}
\alias{mapscale}
\title{A scale function to add beside a map}
\usage{
mapscale(x, y, relwidth = 0.15, metric = TRUE, ratio = TRUE,
  cols = rep(c("black", "white"), 10), cex = 0.8, projected = FALSE,
  epsg = 4326, inset = 0, mark = "Albers Equal Area Conic USGS version",
  madj = 0, ...)
}
\arguments{
\item{x}{A geo-object,}

\item{y}{a state abbrebriation, e.g. "CA",}

\item{relwidth}{margin added to maps, %}

\item{metric}{if metric system is used? TRUE default}

\item{ratio}{default = TRUE}

\item{cols}{black and white interval}

\item{projected}{if projection is used}

\item{inset}{is for box into the plot margin}

\item{mark}{mark is the text line under the scale ba}

\item{madj}{adjustment}

\item{mag}{magnification default =1}
}
\description{
Map scale was modified from the map package add scale to unprojected map
if map has been projected, then use scale bar function, need to specify magnification for that.
see Tanimura et al: auxiliary cartographic functions in R: North Arrow, Scale Bar, and label with a leader arrow
the default ppi if not specified is 72 ppi/inch
if px is specified, decimal degree is used to calculate actual distance
}
\examples{
map("usa") 
mapscale(x="bottomleft", relwidth = 0.2, mark ="", projected = F, ratio=F, inset = c(0.05, -0.05)) 
par(xpd =TRUE)
 plot(0:1,0:1)
 map("state", "ohio")
 mapscale(x="topleft", relwidth = 0.2,ratio=T) 
 mapscale(x="topright", relwidth = 0.2,ratio=F)     
 mapscale(x="bottomright", relwidth = 0.2,ratio=F) 
 mapscale(x="bottomleft", relwidth = 0.2, projected =T, ratio=F)      
 mapscale(x="center", relwidth = 0.2, projected =T, ratio=F)   
 box()
  plot(us.states)
 mapscale(x="bottomright", relwidth = 0.2,ratio=F, projected =T, inset = c(0.1, 0.2)) 
 mapscale(x="topright", relwidth = 0.2, ratio=F, projected =T, epsg = 2163, inset = c(0.1, 0.2))
}
